+++ sdk/src/main/java/com/hedera/hashgraph/sdk/ManagedNode.java
@@ -2,9 +2,7 @@ package com.hedera.hashgraph.sdk;
 
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
-import org.threeten.bp.Duration;
 
-import java.time.Instant;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
 
@@ -22,7 +20,7 @@ abstract class ManagedNode {
 
     void inUse() {
         useCount++;
-        lastUsed = Instant.now().toEpochMilli();
+        lastUsed = System.currentTimeMillis();
     }
 
     synchronized ManagedChannel getChannel() {