+++ CHANGELOG.md
@@ -4,14 +4,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v2.0.5-beta4
+## v2.0.5-beta.5
+
+### Added
+
+ * `TransactionRecord.scheduleRef` - Reference to the scheduled transaction
+ * `TransactionReceipt.scheduledTransactionId`
+ * `ScheduleInfo.scheduledTransactionId`
+ * Feature to copy `TransactionId` of a transaction being scheduled
+   to the parent `ScheduleCreateTransaction` if one is set.
+
+### Fixed
+
+ * `TransactionId.toBytes()` should support `nonce` if set
+ * `TransactionId.fromBytes()` should support `nonce` if set
+
+## v2.0.5-beta.4
 
 ### Added
 
  * Support `memo` for Tokens, Accounts, and Files.
  * `TransactionId.fromString()` should support nonce and scheduled.
 
-## v2.0.5-beta3
+## v2.0.5-beta.3
 
 ### Changed
 
@@ -33,12 +48,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  * `new TransactionId(AccountId, Instant)` - Use `TransactionId.withValidStart()` instead.
 
-## v2.0.5-beta2
+## v2.0.5-beta.2
 
 ### Fixed
  * `Schedule[Create|Sign]Transaction.addScheduleSignature()` didn't save added signatures correctly.
 
-## v2.0.5-beta1
+## v2.0.5-beta.1
 
 ### Added